+++ settings.gradle.kts
@@ -40,3 +40,5 @@ include("kronos-logging")
 include("kronos-compiler-plugin")
 include("kronos-jvm-driver-wrapper")
 include("kronos-testing")
+include("kronos-gradle-plugin")
+include("kronos-maven-plugin")
